+++ src/yafowil/loader.py
@@ -21,6 +21,7 @@ def register():  # noqa
     import yafowil.select
     import yafowil.table
     import yafowil.tag
+    import yafowil.textarea
     import yafowil.url
