+++ src/yafowil/tests/__init__.py
@@ -18,6 +18,7 @@ from yafowil import search
 from yafowil import select
 from yafowil import table
 from yafowil import tag as tag_module
+from yafowil import textarea
 from yafowil import url
 from yafowil.base import factory
 from yafowil.utils import Tag
@@ -49,6 +50,7 @@ class YafowilTestCase(NodeTestCase):
         reload(select)
         reload(table)
         reload(tag_module)
+        reload(textarea)
         reload(url)
 
     def tearDown(self):
@@ -89,6 +91,7 @@ def test_suite():
     from yafowil.tests import test_select
     from yafowil.tests import test_table
     from yafowil.tests import test_tag
+    from yafowil.tests import test_textarea
     from yafowil.tests import test_tsf
     from yafowil.tests import test_url
     from yafowil.tests import test_utils
@@ -115,6 +118,7 @@ def test_suite():
     suite.addTest(unittest.findTestCases(test_select))
     suite.addTest(unittest.findTestCases(test_table))
     suite.addTest(unittest.findTestCases(test_tag))
+    suite.addTest(unittest.findTestCases(test_textarea))
     suite.addTest(unittest.findTestCases(test_tsf))
     suite.addTest(unittest.findTestCases(test_url))
     suite.addTest(unittest.findTestCases(test_utils))
