+++ src/yafowil/tests/test_common.py
@@ -9,10 +9,6 @@ from yafowil.tests import wrapped_fxml
 from yafowil.utils import EMPTY_VALUE
 
 
-###############################################################################
-# Tests
-###############################################################################
-
 class TestCommon(YafowilTestCase):
 
     def test_BC_imports(self):
@@ -76,6 +72,11 @@ class TestCommon(YafowilTestCase):
 
         # tag
         from yafowil.common import tag_renderer
+
+        # textarea
+        from yafowil.common import textarea_managed_props
+        from yafowil.common import textarea_attributes
+        from yafowil.common import textarea_renderer
 
         # url
         from yafowil.common import url_extractor
@@ -476,68 +477,6 @@ class TestCommon(YafowilTestCase):
             value='lorem ipsum',
             mode='skip')
         self.assertEqual(widget(), '')
-
-    def test_textarea_blueprint(self):
-        # Textarea widget
-        widget = factory(
-            'textarea',
-            name='MYTEXTAREA',
-            value=None)
-        self.assertEqual(widget(), (
-            '<textarea class="textarea" cols="80" id="input-MYTEXTAREA" '
-            'name="MYTEXTAREA" rows="25"></textarea>'
-        ))
-
-        widget = factory(
-            'textarea',
-            name='MYTEXTAREA',
-            value=None,
-            props={
-                'data': {
-                    'foo': 'bar'
-                },
-            })
-        self.assertEqual(widget(), (
-            '<textarea class="textarea" cols="80" data-foo=\'bar\' '
-            'id="input-MYTEXTAREA" name="MYTEXTAREA" rows="25"></textarea>'
-        ))
-
-        widget.mode = 'display'
-        self.assertEqual(widget(), (
-            '<div class="display-textarea" data-foo=\'bar\' '
-            'id="display-MYTEXTAREA"></div>'
-        ))
-
-        # Emptyvalue
-        widget = factory(
-            'textarea',
-            name='MYTEXTAREA',
-            props={
-                'emptyvalue': 'EMPTYVALUE',
-            })
-        data = widget.extract(request={'MYTEXTAREA': ''})
-        self.assertEqual(data.name, 'MYTEXTAREA')
-        self.assertEqual(data.value, UNSET)
-        self.assertEqual(data.extracted, 'EMPTYVALUE')
-        self.assertEqual(data.errors, [])
-
-        data = widget.extract(request={'MYTEXTAREA': 'NOEMPTY'})
-        self.assertEqual(data.name, 'MYTEXTAREA')
-        self.assertEqual(data.value, UNSET)
-        self.assertEqual(data.extracted, 'NOEMPTY')
-        self.assertEqual(data.errors, [])
-
-        # Persist
-        widget = factory(
-            'textarea',
-            name='MYTEXTAREA',
-            props={
-                'persist_writer': write_mapping_writer
-            })
-        data = widget.extract(request={'MYTEXTAREA': 'Text'})
-        model = dict()
-        data.write(model)
-        self.assertEqual(model, {'MYTEXTAREA': 'Text'})
 
     def test_bytes_datatype_rendering_and_extraction(self):
         widget = factory(
